% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnc_sp_codeset_file}
\alias{cnc_sp_codeset_file}
\title{Clinical Events & Specialties Codeset File Sample}
\format{
\subsection{\code{cnc_sp_codeset_file}}{

A data frame with 5 columns
\describe{
\item{domain}{A string label for the domain of interest}
\item{domain_tbl}{The name of the CDM table where the variable can be found}
\item{concept_field}{The field in the default_tbl that should be used to join to the codeset}
\item{date_field}{The date field in the default_tbl that should be used to filter the dataset to the cohort period and for longitudinal analyses}
\item{vocabulary_field}{(PCORnet only) The name of the column in the domain table where the vocabulary type is stored}
\item{codeset_name}{The name of the codeset as found in the specs directory; file extension should not be included}
}
}
}
\usage{
cnc_sp_codeset_file
}
\description{
A sample version of the file structure expected for the \code{codeset_tbl}
parameter in the \code{cnc_sp_process} function. The user should recreate
this file and include their own domain definitions.
}
\details{
The codeset listed in the \code{codeset_name} column should be kept in the
\code{file_subdirectory} indicated in \code{initialize_dq_session}
}
\keyword{datasets}
