% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc_sp_output_subfuncs.R
\name{cnc_sp_ms_anom_at}
\alias{cnc_sp_ms_anom_at}
\title{\strong{Multi-Site Across Time Anomaly}
Function to generate output displaying the Euclidean distance between two time series}
\usage{
cnc_sp_ms_anom_at(process_output, grp_vars, specialty_filter = NULL)
}
\arguments{
\item{process_output}{output from \code{cnc_sp_process}
should contain the columns specified in \code{grp_vars}}

\item{grp_vars}{vector of variables used for grouping when identifying all-site summary statistics}

\item{specialty_filter}{a string indicating the SINGLE specialty of intereset that should be displayed at one time}
}
\value{
three graphs:
\enumerate{
\item line graph that shows the smoothed proportion of visits with a
specialty across time computation with the Euclidean distance associated with each line
\item line graph that shows the raw proportion of visits with a
specialty across time computation with the Euclidean distance associated with each line
\item a bar graph with the Euclidean distance value for each site, with the average
proportion as the fill
}
}
\description{
\strong{Multi-Site Across Time Anomaly}
Function to generate output displaying the Euclidean distance between two time series
}
