% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc_sp_output_subfuncs.R
\name{cnc_sp_ms_anom_cs}
\alias{cnc_sp_ms_anom_cs}
\title{\emph{Multi-Site, Anomaly, Cross-Sectional}}
\usage{
cnc_sp_ms_anom_cs(process_output, title, text_wrapping_char = 60)
}
\arguments{
\item{process_output}{the output from the cnc_sp check, summarized to be used in the multi site anomaly detection check}

\item{title}{text containing the title for the plot}

\item{text_wrapping_char}{the number of characters for the specialty names to be displayed on the plot}
}
\value{
a dot plot where the shape of the dot represents whether the point is
anomalous, the color of the dot represents the proportion of visits
for a given specialty, and the size of the dot represents the mean proportion
across all sites
}
\description{
\emph{Multi-Site, Anomaly, Cross-Sectional}
}
