% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc_sp_output_subfuncs.R
\name{cnc_sp_ms_exp_la}
\alias{cnc_sp_ms_exp_la}
\title{\itemize{
\item Multi-Site, Exploratory, Longitudinal*
Function to produce output for clinical event concordance with specialty
for multi site, exploratory, across time
}}
\usage{
cnc_sp_ms_exp_la(data_tbl, facet = NULL)
}
\arguments{
\item{data_tbl}{table which must contain the cols: time_start | codeset_name | specialty_name | site}

\item{facet}{if supplied, variable to facet the plot by}
}
\value{
line plot, with time on x axis, proportion on y, line color determined by site
with a dotted line for the all-site mean
}
\description{
\itemize{
\item Multi-Site, Exploratory, Longitudinal*
Function to produce output for clinical event concordance with specialty
for multi site, exploratory, across time
}
}
