% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc_sp_output.R
\name{cnc_sp_output}
\alias{cnc_sp_output}
\title{Concordance: Clinical Events and Specialties -- Output}
\usage{
cnc_sp_output(
  cnc_sp_process_output,
  cnc_sp_process_names,
  multi_or_single_site,
  anomaly_or_exploratory,
  time = FALSE,
  facet_vars,
  top_n = 15,
  n_mad = 3L,
  specialty_filter = NULL,
  p_value = 0.9
)
}
\arguments{
\item{cnc_sp_process_output}{output from the \code{cnc_sp_process} function}

\item{cnc_sp_process_names}{classified names from the output from the \code{cnc_sp_process} function,
expected to minimally contain the columns:
- specialty_concept_id: unchanged from the \code{conc_process} output
- specialty_name: the assigned classification
this table will be joined to \code{conc_process_output}, so all specialty_concept_id in the \code{conc_process_output} should be in \code{conc_process_names}}

\item{multi_or_single_site}{string indicating whether to generate output for single site or multiple site checks:
- 'multi' for multiple
- 'single' for single}

\item{anomaly_or_exploratory}{string indicating whether to generate output for anomaly detection or exploratory analysis:
- 'anomaly' for anomaly detection
- 'exploratory' for exploratory analysis}

\item{time}{TRUE if should have a time dimension}

\item{facet_vars}{vector of variable names to facet by}

\item{top_n}{integer value for choosing the "top n" to display per check, with meaning dependent on the context of the check}

\item{n_mad}{number of MAD from the median for which to flag anomalies
defaults to 3}

\item{specialty_filter}{an optional filter to apply to the specialty_name field to narrow down results}

\item{p_value}{the p value to be used as a threshold in the multi-site anomaly detection analysis}
}
\value{
the corresponding visualization/s for the site level (multi/single), time dimension,
and analysis level (exploratory/anomaly detection) specified
}
\description{
Concordance: Clinical Events and Specialties -- Output
}
