% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnc_sp_specialty_names}
\alias{cnc_sp_specialty_names}
\title{Clinical Events & Specialties - Specialty Naming Sample}
\format{
\subsection{\code{cnc_sp_specialty_names}}{

A data frame with 5 rows and 3 columns
\describe{
\item{specialty_concept_id}{The specialty_concept_id identified in \code{cnc_sp_process}}
\item{specialty_concept_name}{If a vocab_tbl was provided, the concept_name associated with the specialty_concept_id; otherwise will default to "No vocabulary table input"}
\item{specialty_name}{A user-provided identifier for the specialty; can be repeated across rows to group specialties together}
}
}
}
\usage{
cnc_sp_specialty_names
}
\description{
A sample version of how the list of specialties output by \code{cnc_sp_process} should be
labelled for use in the \code{cnc_sp_output} function.
}
\keyword{datasets}
