% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc_sp_output_subfuncs.R
\name{cnc_sp_ss_anom_cs}
\alias{cnc_sp_ss_anom_cs}
\title{\emph{Single Site, Anomaly, Cross-Sectional}}
\usage{
cnc_sp_ss_anom_cs(data_tbl, facet = NULL)
}
\arguments{
\item{data_tbl}{table which must contain the cols: specialty_name | prop | median | n_mad}

\item{facet}{vector of variable names to be used to facet the graph}
}
\value{
a dot plot where the shape of the dot represents whether the point is
anomalous, the color of the dot represents the proportion of visits
for a given cluster, and the size of the dot represents the mean proportion
across all specialties
}
\description{
Function to produce output for clinical event concordance with specialty
for single site, anomaly, not over time
}
