% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc_sp_output_subfuncs.R
\name{cnc_sp_ss_exp_cs}
\alias{cnc_sp_ss_exp_cs}
\title{\itemize{
\item Single Site, Exploratory, Cross-Sectional*
Function to produce output for clinical event concordance with specialty
for single site, exploratory, not over time
}}
\usage{
cnc_sp_ss_exp_cs(data_tbl, facet, x_var, y_var, fill_var, top_n)
}
\arguments{
\item{data_tbl}{table with the data to plot}

\item{facet}{list of one or more variables to facet the plot on}

\item{x_var}{variable to plot on the x axis}

\item{y_var}{variable to plot on the y axis}

\item{fill_var}{variable to fill bars with}

\item{top_n}{an integer indicating the top N number of specialties to display on the graph}
}
\value{
a bar plot based on the values of \code{x_var}, \code{y_var}, \code{fill_var}, \code{facet}
}
\description{
\itemize{
\item Single Site, Exploratory, Cross-Sectional*
Function to produce output for clinical event concordance with specialty
for single site, exploratory, not over time
}
}
\keyword{internal}
