% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc_sp_output_subfuncs.R
\name{cnc_sp_ss_exp_la}
\alias{cnc_sp_ss_exp_la}
\title{\emph{Single Site, Exploratory, Longitudinal}
Function to produce output for clinical event concordance with specialty
for single site, exploratory, across time}
\usage{
cnc_sp_ss_exp_la(data_tbl, facet = NULL)
}
\arguments{
\item{data_tbl}{table with the data to plot}

\item{facet}{list of one or more variables to facet the plot on}
}
\value{
a plotly line plot of the proportion of visits with each specialty
against time, with line color representing specialty
}
\description{
\emph{Single Site, Exploratory, Longitudinal}
Function to produce output for clinical event concordance with specialty
for single site, exploratory, across time
}
