% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnc_sp_visit_file_omop}
\alias{cnc_sp_visit_file_omop}
\title{Clinical Facts & Specialties Visit File Sample -- OMOP version}
\format{
\subsection{cnc_sp_visit_file_omop}{

A data frame with 2 columns and 8 rows
\describe{
\item{visit_concept_id}{The visit_concept_id as it appears in the visit_occurrence table}
\item{visit_type}{A string to label the visit type of the visit_concept_id; this string is what should be referenced in the \code{visit_types} parameter}
}
}
}
\usage{
cnc_sp_visit_file_omop
}
\description{
A sample version of the file structure expected for the \code{visit_type_tbl}
parameter in the \code{cnc_sp_process} function. The user should recreate
this file and include their own domain definitions.
}
\keyword{datasets}
