% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnc_sp_visit_file_pcornet}
\alias{cnc_sp_visit_file_pcornet}
\title{Clinical Events & Specialties Visit File Sample -- PCORnet version}
\format{
\subsection{cnc_sp_visit_file_pcornet}{

A data frame with 2 columns
\describe{
\item{enc_type}{The enc_type as it appears in the encounter table}
\item{visit_type}{A string to label the visit type of the enc_type}
}
}
}
\usage{
cnc_sp_visit_file_pcornet
}
\description{
A sample version of the file structure expected for the \code{visit_type_tbl}
parameter in the \code{cnc_sp_process} function. The user should recreate
this file and include their own domain definitions.
}
\keyword{datasets}
