% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc_sp_output_subfuncs.R
\name{insert_top_n_indicator}
\alias{insert_top_n_indicator}
\title{Function to insert a column into a table indicating whether the record is in the top n
for the given group}
\usage{
insert_top_n_indicator(dat, gp_cols, val_col, n, sum_first = FALSE)
}
\arguments{
\item{dat}{table containing the data}

\item{gp_cols}{columns to group by
the top n will be taken from each group, so the total indicated as "top" hits will be # of groups * n}

\item{val_col}{column to order by when determining the top n}

\item{n}{number of records to indicate as top within the group}

\item{sum_first}{boolean indicating whether a value should be grouped and summed prior to identifying the top N}
}
\value{
the original \code{dat} table with all original columns,
plus a column \code{top_n_indicator} which is TRUE if the record is in the top n
and FALSE if not
}
\description{
Function to insert a column into a table indicating whether the record is in the top n
for the given group
}
