% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc_sp_output_subfuncs.R
\name{cnc_sp_ms_exp_cs}
\alias{cnc_sp_ms_exp_cs}
\title{\emph{Multi-Site, Exploratory, Cross-Sectional}
Function to produce output for clinical event concordance with specialty
for multi site, exploratory, not over time}
\usage{
cnc_sp_ms_exp_cs(data_tbl, facet = NULL)
}
\arguments{
\item{data_tbl}{table with the data to plot}

\item{facet}{list of one or more variables to facet the plot on}
}
\value{
a dot plot of specialty against proportion of visits with that specialty
at each site, with dot color representing site
}
\description{
\emph{Multi-Site, Exploratory, Cross-Sectional}
Function to produce output for clinical event concordance with specialty
for multi site, exploratory, not over time
}
