% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc_sp_output.R
\name{cnc_sp_output}
\alias{cnc_sp_output}
\title{Clinical Events and Specialties -- Output Generation}
\usage{
cnc_sp_output(
  cnc_sp_process_output,
  cnc_sp_process_names,
  facet_vars = NULL,
  top_n = 15,
  n_mad = 3L,
  specialty_filter = NULL,
  p_value = 0.9,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{cnc_sp_process_output}{\emph{tabular input} || \strong{required}

The tabular output (with visit-based counts per specialty) produced by \code{cnc_sp_process}}

\item{cnc_sp_process_names}{\emph{tabular input} || \strong{required}

The tabular output (with specialty names & any specialty_name grouping categories) produced by \code{cnc_sp_process}

To see an example of what this file should look like, see \code{?clinicalevents.specialties::cnc_sp_specialty_names}}

\item{facet_vars}{\emph{string or vector} || defaults to \code{NULL}

A string or vector representing the variables by which the plot should be facet.
Accepted values are \code{cluster} and/or \code{visit_type}}

\item{top_n}{\emph{integer} || defaults to \code{15}

An integer value indicating the cutoff for the top N of each group to display per check}

\item{n_mad}{\emph{integer} || defaults to \code{3}

An integer indicating the number of MAD from the median that should be
considered the threshold for an anomalous value}

\item{specialty_filter}{\emph{string or vector} || defaults to \code{NULL}

An optional parameter indicating the specialty or specialties to limit to in the analysis}

\item{p_value}{\emph{numeric} || defaults to \code{0.9}

The p value to be used as a threshold in the Multi-Site,
Anomaly Detection, Cross-Sectional analysis}

\item{large_n}{\emph{boolean} || defaults to \code{FALSE}

For Multi-Site analyses, a boolean indicating whether the large N
visualization, intended for a high volume of sites, should be used. This
visualization will produce high level summaries across all sites, with an
option to add specific site comparators via the \code{large_n_sites} parameter.}

\item{large_n_sites}{\emph{vector} || defaults to \code{NULL}

When \code{large_n = TRUE}, a vector of site names that can add site-level information
to the plot for comparison across the high level summary information.}
}
\value{
This function will produce a graph to visualize the results
from \code{cnc_sp_process} based on the parameters provided. The default
output is typically a static ggplot or gt object, but interactive
elements can be activated by passing the plot through \code{make_interactive_squba}.
For a more detailed description of output specific to each check type,
see the PEDSpace metadata repository
}
\description{
Using BOTH tabular outputs generated by \code{cnc_sp_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_squba()}
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'clinicalevents.specialties'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'cnc_sp_process_test',
                      working_directory = my_directory,
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = my_file_folder,
                      cdm_schema = NA)

## Turn off SQL trace for this example
config('db_trace', FALSE)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000),
                #RSQLite does not store date objects,
                #hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Prepare input tables
cnc_sp_visit_tbl <- dplyr::tibble(visit_concept_id = c(9201,9202,9203),
                                  visit_type = c('inpatient', 'outpatient', 'emergency'))

cnc_sp_concept_tbl <- dplyr::tibble(domain = 'Hypertension',
                                    domain_tbl = 'condition_occurrence',
                                    concept_field = 'condition_concept_id',
                                    date_field = 'condition_start_date',
                                    vocabulary_field = NA,
                                    codeset_name = 'dx_hypertension')

#' Execute `cnc_sp_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
cnc_sp_process_example <- cnc_sp_process(cohort = cohort,
                                         omop_or_pcornet = 'omop',
                                         multi_or_single_site = 'single',
                                         anomaly_or_exploratory = 'exploratory',
                                         codeset_tbl = cnc_sp_concept_tbl,
                                         visit_type_tbl = cnc_sp_visit_tbl,
                                         time = FALSE) \%>\%
  suppressMessages()

cnc_sp_process_example$cnc_sp_process_output
cnc_sp_process_example$cnc_sp_process_names

#' Execute `cnc_sp_output` function
cnc_sp_output_example <-
  cnc_sp_output(cnc_sp_process_output =
                  cnc_sp_process_example$cnc_sp_process_output,
                cnc_sp_process_names =
                  cnc_sp_process_example$cnc_sp_process_names \%>\%
                    dplyr::mutate(specialty_name = 'General Pediatrics'),
                facet_vars = c('visit_type')) \%>\%
  suppressMessages()

cnc_sp_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(cnc_sp_output_example)
}
