% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc_sp_process.R
\name{cnc_sp_process}
\alias{cnc_sp_process}
\title{Clinical Events and Specialties}
\usage{
cnc_sp_process(
  cohort,
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  omop_or_pcornet,
  age_groups = NULL,
  codeset_tbl,
  care_site = FALSE,
  provider = TRUE,
  visit_detail = FALSE,
  visit_type_tbl = NULL,
  time = FALSE,
  time_span = c("2012-01-01", "2020-01-01"),
  time_period = "year",
  vocab_tbl = NULL
)
}
\arguments{
\item{cohort}{\emph{tabular input} || \strong{required}

The cohort to be used for data quality testing. This table should contain,
at minimum:
\itemize{
\item \code{site} | \emph{character} | the name(s) of institutions included in your cohort
\item \code{person_id} / \code{patid} | \emph{integer} / \emph{character} | the patient identifier
\item \code{start_date} | \emph{date} | the start of the cohort period
\item \code{end_date} | \emph{date} | the end of the cohort period
}

Note that the start and end dates included in this table will be used to
limit the search window for the analyses in this module.}

\item{multi_or_single_site}{\emph{string} || defaults to \code{single}

A string, either \code{single} or \code{multi}, indicating whether a single-site or
multi-site analysis should be executed}

\item{anomaly_or_exploratory}{\emph{string} || defaults to \code{exploratory}

A string, either \code{anomaly} or \code{exploratory}, indicating what type of results
should be produced.

Exploratory analyses give a high level summary of the data to examine the
fact representation within the cohort. Anomaly detection analyses are
specialized to identify outliers within the cohort.}

\item{omop_or_pcornet}{\emph{string} || \strong{required}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data
\itemize{
\item \code{omop}: run the \code{\link[=cnc_sp_process_omop]{cnc_sp_process_omop()}} function against an OMOP CDM instance
\item \code{pcornet}: run the \code{\link[=cnc_sp_process_pcornet]{cnc_sp_process_pcornet()}} function against a PCORnet CDM instance
}}

\item{age_groups}{\emph{tabular input} || defaults to \code{NULL}

If you would like to stratify the results by age group, create a table or
CSV file with the following columns and use it as input to this parameter:
\itemize{
\item \code{min_age} | \emph{integer} | the minimum age for the group (i.e. 10)
\item \code{max_age} | \emph{integer} | the maximum age for the group (i.e. 20)
\item \code{group} | \emph{character} | a string label for the group (i.e. 10-20, Young Adult, etc.)
}

If you would \emph{not} like to stratify by age group, leave as \code{NULL}}

\item{codeset_tbl}{\emph{tabular input} || \strong{required}

A table defining the clinical event of interest, containing the following:
\itemize{
\item \code{domain} | \emph{character} | a string label for the domain where the event is defined
\item \code{domain_tbl} | \emph{character} | the name of the CDM table where the event is defined
\item \code{concept_field} | \emph{character} | the string name of the field in the domain table where the concepts are located
\item \code{date_field} | \emph{character} | the name of the field in the domain table with the date that should be used for temporal filtering
\item \code{codeset_name} | \emph{character} | name of the codeset with concepts defining the clinical event
}

The codeset file identified by this table can optionally contain a \code{cluster} column
specifying subgroups of the codeset, and if so, the results will be stratified by cluster

To see an example of the structure of this file, see \code{?clinicalevents.specialties::cnc_sp_codeset_file}}

\item{care_site}{\emph{boolean} || defaults to \code{FALSE}

A boolean indicating whether care site/facility specialty values should be included
in the analysis. If both \code{provider} and \code{care_site} are TRUE, provider specialty will
be prioritized.}

\item{provider}{\emph{boolean} | defaults to \code{TRUE}

A boolean indicating whether care provider specialty values should be included
in the analysis. If both \code{provider} and \code{care_site} are TRUE, provider specialty will
be prioritized.}

\item{visit_detail}{\emph{boolean} || defaults to \code{FALSE}

For OMOP analyses only -- a boolean indicating whether the visit_detail table should
be used as the primary visit table to identify specialty visits. If left FALSE, visit_occurrence
will be used.}

\item{visit_type_tbl}{\emph{tabular input} || defaults to \code{NULL}

A table defining available visit types to be used as an optional stratification.
This table should contain the following field:
\itemize{
\item \code{visit_concept_id} / \code{visit_detail_concept_id} or \code{enc_type} | \emph{integer} / \emph{character} | the visit type identifier that represents the visit type of interest (ex: 9201 or IP)
\item \code{visit_type} | \emph{character} | a string description of the visit type
}

To see an example of the structure of this file, see \code{?clinicalevents.specialties::cnc_sp_visit_file_omop} or
\code{?clinicalevents.specialties::cnc_sp_visit_file_pcornet}}

\item{time}{\emph{boolean} || defaults to \code{FALSE}

A boolean to indicate whether to execute a longitudinal analysis}

\item{time_span}{\emph{vector - length 2} || defaults to \code{c('2012-01-01', '2020-01-01')}

A vector indicating the lower and upper bounds of the time series for longitudinal analyses}

\item{time_period}{\emph{string} || defaults to \code{year}

A string indicating the distance between dates within the specified time_span.
Defaults to \code{year}, but other time periods such as \code{month} or \code{week} are
also acceptable}

\item{vocab_tbl}{\emph{tabular input} || defaults to \code{NULL}

A vocabulary table containing concept names that will be used to retrieve labels
for the specialty concepts (ex: the OMOP concept table)}
}
\value{
This function will return two dataframes:
\itemize{
\item A table containing all of the specialties associated with the clinical event, which can be further grouped by the user before feeding this into the cnc_sp_output function
\item A table containing counts of visits, optionally stratified by visit, cluster, and/or time period, with each specialty for the visits meeting criteria (i.e. those with the clinical fact provided)
}

For a more detailed description of output specific to each check type, see the PEDSpace metadata repository
}
\description{
This is a concordance module that will assess the types of specialty care received by and
quality of specialty data found in a study sample. The user will provide a
clinical codeset of interest (\code{codeset_tbl}) with an associated domain
and will be able to stratify results by: visit type
(with user-provided groupings in \code{visit_type_tbl}), cluster
(an additional column added to \code{codeset_tbl} with subgroupings), or time
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'clinicalevents.specialties'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'cnc_sp_process_test',
                      working_directory = my_directory,
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = my_file_folder,
                      cdm_schema = NA)

## Turn off SQL trace for this example
config('db_trace', FALSE)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000),
                #RSQLite does not store date objects,
                #hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Prepare input tables
cnc_sp_visit_tbl <- dplyr::tibble(visit_concept_id = c(9201,9202,9203),
                                  visit_type = c('inpatient', 'outpatient', 'emergency'))

cnc_sp_concept_tbl <- dplyr::tibble(domain = 'Hypertension',
                                    domain_tbl = 'condition_occurrence',
                                    concept_field = 'condition_concept_id',
                                    date_field = 'condition_start_date',
                                    vocabulary_field = NA,
                                    codeset_name = 'dx_hypertension')

#' Execute `cnc_sp_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
cnc_sp_process_example <- cnc_sp_process(cohort = cohort,
                                         omop_or_pcornet = 'omop',
                                         multi_or_single_site = 'single',
                                         anomaly_or_exploratory = 'exploratory',
                                         codeset_tbl = cnc_sp_concept_tbl,
                                         visit_type_tbl = cnc_sp_visit_tbl,
                                         time = FALSE) \%>\%
  suppressMessages()

cnc_sp_process_example$cnc_sp_process_output
cnc_sp_process_example$cnc_sp_process_names

#' Execute `cnc_sp_output` function
cnc_sp_output_example <-
  cnc_sp_output(cnc_sp_process_output =
                  cnc_sp_process_example$cnc_sp_process_output,
                cnc_sp_process_names =
                  cnc_sp_process_example$cnc_sp_process_names \%>\%
                    dplyr::mutate(specialty_name = 'General Pediatrics'),
                facet_vars = c('visit_type')) \%>\%
  suppressMessages()

cnc_sp_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(cnc_sp_output_example)
}
