% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc_sp_process.R
\name{cnc_sp_process}
\alias{cnc_sp_process}
\title{Concordance: Clinical Events and Specialties}
\usage{
cnc_sp_process(
  cohort,
  multi_or_single_site = "multi",
  omop_or_pcornet,
  age_groups = NULL,
  codeset_tbl,
  care_site = FALSE,
  provider = TRUE,
  visit_type_tbl = NULL,
  time = FALSE,
  time_span = c("2012-01-01", "2020-01-01"),
  time_period = "year",
  vocab_tbl = NULL
)
}
\arguments{
\item{cohort}{\itemize{
\item A dataframe with the cohort of patients for your study. Should include the columns:
\code{site} | \code{person_id} | \code{start_date} | \code{end_date}
}}

\item{multi_or_single_site}{Option to run the function on a single vs multiple sites
- \code{single}: run on a single site, or treat all of the sites as one
- \code{multi}: run on a group of sites, treating each site separately}

\item{omop_or_pcornet}{Option to run the function using the OMOP or PCORnet CDM as the default CDM
\itemize{
\item \code{omop}: run the \code{\link[=cnc_sp_process_omop]{cnc_sp_process_omop()}} function against an OMOP CDM instance
\item \code{pcornet}: run the \code{\link[=cnc_sp_process_pcornet]{cnc_sp_process_pcornet()}} function against a PCORnet CDM instance
}}

\item{age_groups}{If you would like to stratify the results by age group,  create a table or CSV file with the following
columns and include it as the \code{age_groups} function parameter:
\itemize{
\item \code{min_age}: the minimum age for the group (i.e. 10)
\item \code{max_age}: the maximum age for the group (i.e. 20)
\item \code{group}: a string label for the group (i.e. 10-20, Young Adult, etc.)
}

If you would \emph{not} like to stratify by age group, leave the argument as NULL}

\item{codeset_tbl}{table in the specs directory with the columns:
\itemize{
\item domain: name of the domain
\item domain_tbl: name of the cdm_tbl
\item concept_field: column name in the domain_tbl for which to search the codeset concept_ids
\item date_field: column name in the domain_tbl to be used for time-based filtering
\item codeset_name: name of a codeset that exists as a csv file in the specs directory. The codeset can optionally contain a \code{cluster} column specifying subgroups of the codeset, and if so, the results will be stratified by cluster
}}

\item{care_site}{TRUE if want to look at care_site specialty
FALSE if do not want to look at care_site specialty}

\item{provider}{TRUE if want to look at provider specialty
FALSE if do not want to look at provider specialty
IF both \code{provider} and \code{care_site} are both TRUE,
provider specialty will be prioritized if provider and care_site are discordant for the visit}

\item{visit_type_tbl}{\itemize{
\item a csv file that defines available visit types that are called in \code{visit_types.} defaults to the provided
\code{conc_visit_types.csv} file, which contains the following fields:
- visit_concept_id: the visit_concept_id that represents the visit type of interest (i.e. 9201)
- visit_type: the string label to describe the visit type; this label can be used multiple times
within the file if multiple visit_concept_ids represent the visit type

\if{html}{\out{<div class="sourceCode">}}\preformatted{                    This CSV can be altered to fit the users needs, or another csv with the same columns and formatting can be supplied.
}\if{html}{\out{</div>}}
}}

\item{time}{TRUE if results should be over time. Defaults to FALSE}

\item{time_span}{if time=TRUE, vector containing minimum and maximum dates over which to measure}

\item{time_period}{if time=TRUE, indicates time period (e.g. 'year', 'month') over which to measure}

\item{vocab_tbl}{location of vocabulary table containing concept_id to concept_name mapping. If a vocabulary table is not available, will default to NULL}
}
\value{
2 tables:
1 table stored as a csv file in the \code{results} directory containing all of the specialties in the results of the
DQ check, with the columns:
- specialty_concept_id: an identifier for the specialty based on the data model
- specialty_concept_name: if a \code{vocab_tbl} is provided, the name of the specialty
that corresponds to each specialty_concept_id. If no \code{vocab_tbl} is provided,
defaults to 'No vocabulary table input'
1 table containing counts of visits, optionally stratified by visit and/or time period,
with each specialty for the visits meeting criteria (i.e. those with the clinical fact provided)
}
\description{
This is a check that will assess quality of specialty data in a study sample
The user will provide a clinical codeset of interest with an associated domain
and will be able to stratify results by:
visit type (user will provide groupings)
time
}
