% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc_sp_process_subfuncs_omop.R
\name{find_distinct_concepts}
\alias{find_distinct_concepts}
\title{Function to generate a table of concept_id + concept_name for a set of concepts}
\usage{
find_distinct_concepts(
  tbl,
  concept_col = "concept_id",
  vocab = vocabulary_tbl("concept")
)
}
\arguments{
\item{tbl}{table with a specialty_concept_id column}

\item{concept_col}{the column in the vocabulary table that should be used to join to
the input tbl}

\item{vocab}{table with a concept_id and concept_name column, defaulting to the
vocabulary_tbl \code{concept}}
}
\value{
table with distinct specialty_concept_id | concept_name
}
\description{
Function to generate a table of concept_id + concept_name for a set of concepts
}
